pyproject.toml
[project]
name = "rap-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.1.3",
    "llama-parse>=0.6.94",
    "openai>=2.30.0",
    "pandas>=2.2.0",
    "pyarrow>=18.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.3.0",
    "tiktoken>=0.12.0",
]
